Renaming modules and directories to change gem name
