--- bin/aws-auditor
@@ -1,27 +1,0 @@
-#!/usr/bin/env ruby
-require 'rubygems'
-require 'gli'
-require_relative '../lib/aws_auditor'
-
-include GLI::App
-
-program_desc 'AWS Auditor'
-version AwsAuditor::VERSION
-
-wrap_help_text :verbatim
-
-flag [:config], :desc => 'AwsAuditor config file path', :default_value => AwsAuditor::DefaultPaths.config
-
-program_long_desc """
-DOCUMENTATION
-"""
-
-commands_from File.expand_path(File.dirname(__FILE__) + '/../lib/aws_auditor/commands')
-
-pre do |global,command,options,args|
-  AwsAuditor::Config.load(global[:config])
-  AwsAuditor::Config.merge! global
-  true
-end
-
-exit run(ARGV)
