--- lib/aws_auditor.rb
@@ -1,14 +1,0 @@
-require_relative 'aws_auditor/version'
-require_relative 'aws_auditor/convenience_wrappers'
-require_relative 'aws_auditor/ec2_instance'
-require_relative 'aws_auditor/rds_instance'
-require_relative 'aws_auditor/cache_instance'
-require_relative 'aws_auditor/stack'
-require_relative 'aws_auditor/google_sheet'
-require_relative 'aws_auditor/output'
-require_relative 'aws_auditor/config'
-require_relative 'aws_auditor/notify_slack'
-
-module AwsAuditor
-
-end
