--- lib/aws_auditor/version.rb
@@ -1,3 +1,0 @@
-module AwsAuditor
-  VERSION = "0.2.0"
-end
