+++ lib/sport_ngin_aws_auditor.rb
@@ -1,0 +1,14 @@
+require_relative 'sport_ngin_aws_auditor/version'
+require_relative 'sport_ngin_aws_auditor/convenience_wrappers'
+require_relative 'sport_ngin_aws_auditor/ec2_instance'
+require_relative 'sport_ngin_aws_auditor/rds_instance'
+require_relative 'sport_ngin_aws_auditor/cache_instance'
+require_relative 'sport_ngin_aws_auditor/stack'
+require_relative 'sport_ngin_aws_auditor/google_sheet'
+require_relative 'sport_ngin_aws_auditor/output'
+require_relative 'sport_ngin_aws_auditor/config'
+require_relative 'sport_ngin_aws_auditor/notify_slack'
+
+module SportNginAwsAuditor
+
+end
