+++ lib/sport_ngin_aws_auditor/output.rb
@@ -1,0 +1,13 @@
+require 'highline'
+
+module SportNginAwsAuditor
+  module Output
+    def self.terminal
+      @terminal ||= HighLine.new
+    end
+
+    def self.ask(*args, &block)
+      terminal.ask(*args, &block)
+    end
+  end
+end
