+++ lib/sport_ngin_aws_auditor/version.rb
@@ -1,0 +1,3 @@
+module SportNginAwsAuditor
+  VERSION = "0.2.0"
+end
